+++ marlowe-test/src/Spec/Marlowe/Semantics/Next/Contract/When.hs
@@ -9,9 +9,9 @@ module Spec.Marlowe.Semantics.Next.Contract.When
 
 import Data.Bifunctor (first)
 import Data.List.Index
+import Language.Marlowe.Core.V1.Next.Indexed
+import Language.Marlowe.Core.V1.Next.IsMerkleizedContinuation
 import Language.Marlowe.Core.V1.Semantics (ReduceResult(ContractQuiescent), reduceContractUntilQuiescent)
-import Language.Marlowe.Core.V1.Semantics.Next.Indexed
-import Language.Marlowe.Core.V1.Semantics.Next.IsMerkleizedContinuation
 import Language.Marlowe.Core.V1.Semantics.Types (Action, Case(..), Contract(When), Environment, State, Timeout)
 import Spec.Marlowe.Semantics.Arbitrary ()
