+++ marlowe/src/Language/Marlowe/Core/V1/Next/Applicables/CanNotify.hs
@@ -10,7 +10,7 @@
 {-# LANGUAGE RankNTypes #-}
 
 
-module Language.Marlowe.Core.V1.Semantics.Next.Applicables.CanNotify
+module Language.Marlowe.Core.V1.Next.Applicables.CanNotify
   ( CanNotify(..)
   ) where
 
@@ -20,8 +20,8 @@ import Data.Aeson.Types ()
 import Deriving.Aeson (Generic)
 import Language.Marlowe.Pretty (Pretty(..))
 
-import Language.Marlowe.Core.V1.Semantics.Next.Indexed (CaseIndex(CaseIndex), Indexed(..))
-import Language.Marlowe.Core.V1.Semantics.Next.IsMerkleizedContinuation (IsMerkleizedContinuation)
+import Language.Marlowe.Core.V1.Next.Indexed (CaseIndex(CaseIndex), Indexed(..))
+import Language.Marlowe.Core.V1.Next.IsMerkleizedContinuation (IsMerkleizedContinuation)
 import Prelude
