+++ marlowe/src/Language/Marlowe/Core/V1/Semantics/Types.hs
@@ -67,9 +67,6 @@ module Language.Marlowe.Core.V1.Semantics.Types
   , TokenName(..)
   , Value(..)
   , ValueId(..)
-  , environment
-  , isNotTimedOut
-  , isTimedOut
     -- * Error Types
   , IntervalError(..)
     -- * Utility Functions
@@ -114,9 +111,6 @@ import PlutusTx.Prelude hiding (encodeUtf8, mapM, (<$>), (<*>), (<>))
 import Prelude (mapM, (<$>))
 import qualified Prelude as Haskell
 import Text.PrettyPrint.Leijen (parens, text)
-
-import Data.Time (UTCTime)
-import Plutus.V1.Ledger.SlotConfig (utcTimeToPOSIXTime)
 
 
 -- Functions that used in Plutus Core must be inlinable,
@@ -315,17 +309,6 @@ data State = State { accounts    :: Accounts
 newtype Environment = Environment { timeInterval :: TimeInterval }
   deriving stock (Haskell.Show,Haskell.Eq,Haskell.Ord)
 
-environment :: UTCTime -> UTCTime -> Environment
-environment start end = Environment (utcTimeToPOSIXTime start, utcTimeToPOSIXTime end)
-
-isTimedOut :: POSIXTime -> Environment -> Bool
-isTimedOut timeout Environment {timeInterval = (start,_)} | timeout  <= start = True
-isTimedOut _ _ = False
-
-isNotTimedOut :: POSIXTime -> Environment -> Bool
-isNotTimedOut timeout = not . isTimedOut timeout
-
-
 instance FromJSON Environment where
   parseJSON = withObject "Environment"
     (\v -> Environment <$> (posixIntervalFromJSON =<< v .: "timeInterval"))
@@ -333,8 +316,6 @@ instance FromJSON Environment where
 instance ToJSON Environment where
   toJSON Environment{..} = object
     [ "timeInterval" .= posixIntervalToJSON timeInterval]
-
-
 
 -- | Input for a Marlowe contract. Correspond to expected 'Action's.
 data InputContent = IDeposit AccountId Party Token Integer
